2024-09-17-upd: add folder IO mode

+++ src/main.py
@@ -1,13 +1,14 @@
 import cv2
 from process import *
 
+# params
+debug = False
+size = 4
+filename_prefix = 'demo'
 
-if __name__ == '__main__':
-    id: int = 4
-    img = load_img(f'../assets/input/demo{id}.jpg')
-    background_img = load_img('../assets/input/background.jpg')
-    diff_img = diff_img(img, background_img)
-    gray_img = gbr_to_gray(diff_img)
+def detect_once(img: cv2.Mat, background_img: cv2.Mat, i: int):
+    _diff_img = diff_img(img, background_img)
+    gray_img = gbr_to_gray(_diff_img)
     binary_img = gbr_to_binary(gray_img)
     denoised_img = denoise_img(binary_img)
     contours = find_contour_by_canny(denoised_img)
@@ -17,15 +18,22 @@ if __name__ == '__main__':
     cv2.drawContours(contour_img, contours, -1, (0, 255, 0), 2)
     cv2.drawContours(contour_raw_img, contours, -1, (0, 255, 0), 2)
 
-    # display result
-    cv2.imshow('origin', img)
-    cv2.imshow('gray', gray_img)
-    cv2.imshow('binary', binary_img)
-    cv2.imshow('denoised', denoised_img)
-    cv2.imshow('contours', contour_img)
+    if debug:
+        # display result
+        cv2.imshow('origin', img)
+        cv2.imshow('gray', gray_img)
+        cv2.imshow('binary', binary_img)
+        cv2.imshow('denoised', denoised_img)
+        cv2.imshow('contours', contour_img)
 
-    cv2.imwrite(f'../assets/output/contours{id}.jpg', contour_img)
-    cv2.imwrite(f'../assets/output/contours{id}_on_raw.jpg', contour_raw_img)
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
+    
+    cv2.imwrite(f'../assets/output_dir/contours{i+1}.jpg', contour_img)
+    cv2.imwrite(f'../assets/output_dir/contours{i+1}_on_raw.jpg', contour_raw_img)
 
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+if __name__ == '__main__':
+    for i in range(size):
+        img = load_img(f'../assets/input_dir/{filename_prefix}{i+1}.jpg')
+        background_img = load_img('../assets/input_dir/background.jpg')
+        detect_once(img, background_img, i)
